+++ lib/modules/cache/cache_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: cache
-// Updated at: 2026-04-17 12:57:46
-// Commit: feat(cache): add bi-directional event bus handler for cache
+// Updated at: 2026-04-17 21:53:56
+// Commit: feat(cache): implement shimmer loading skeleton in cache
 
 class cacheComponentState {
-  final String commitRef = '471';
-  final String lastMessage = 'feat(cache): add bi-directional event bus handler for cache';
-  final DateTime timestamp = DateTime.parse('2026-04-17T12:57:46+05:00');
+  final String commitRef = '499';
+  final String lastMessage = 'feat(cache): implement shimmer loading skeleton in cache';
+  final DateTime timestamp = DateTime.parse('2026-04-17T21:53:56+05:00');
 }